message.service & thread.service decoupled
