+++ packages/app/src/messages/message.module.ts
@@ -12,9 +12,8 @@ import { FileSystemService } from "App/file-system/services/file-system.service.
 import { IndexStorage } from "App/index-storage/types"
 import { BaseModule } from "App/core/module"
 import { MessageModel, ThreadModel } from "App/messages/models"
-import { MessageService } from "App/messages/services"
+import { MessageService, ThreadService } from "App/messages/services"
 import { MessageController } from "App/messages/controllers"
-import { ThreadController } from "App/messages/controllers/thread.controller"
 import { MessageObserver } from "App/messages/observers/message.observer"
 import { MessageRepository, ThreadRepository } from "App/messages/repositories"
 
@@ -40,20 +39,21 @@ export class MessageModule extends BaseModule {
 
     const messageModel = new MessageModel(this.index, this.eventEmitter)
     const threadModel = new ThreadModel(this.index, this.eventEmitter)
-    const messageService = new MessageService(this.deviceService)
+    const threadService = new ThreadService(this.deviceService)
+    const messageService = new MessageService(this.deviceService, threadService)
     const messageController = new MessageController(messageService)
-    const threadController = new ThreadController(messageService)
     const messageRepository = new MessageRepository(messageModel)
     const threadRepository = new ThreadRepository(threadModel)
     const messageObserver = new MessageObserver(
       this.ipc,
       this.deviceService,
       messageService,
+      threadService,
       messageRepository,
       threadRepository
     )
     this.models = [messageModel, threadModel]
-    this.controllers = [messageController, threadController]
+    this.controllers = [messageController]
     this.observers = [messageObserver]
   }
 }
