+++ packages/app/src/messages/presenters/message.presenter.ts
@@ -8,10 +8,8 @@ import {
   MessagesCategory as PureMessagesCategory,
   MessageType as PureMessageType,
   PostMessagesBody,
-  Thread as PureThread,
 } from "@mudita/pure"
-import { Message, MessageType, NewMessage, Thread } from "App/messages/reducers"
-import { Feature, flags } from "App/feature-flags"
+import { Message, MessageType, NewMessage } from "App/messages/reducers"
 
 export type AcceptablePureMessageType =
   | PureMessageType.FAILED
@@ -27,24 +25,6 @@ export class MessagePresenter {
       number: newMessage.phoneNumber,
       messageBody: newMessage.content,
       category: PureMessagesCategory.message,
-    }
-  }
-
-  static mapToThread(pureThread: PureThread): Thread {
-    const {
-      isUnread,
-      lastUpdatedAt,
-      messageSnippet,
-      threadID,
-      number = "",
-    } = pureThread
-    return {
-      messageSnippet,
-      // TODO: turn on in https://appnroll.atlassian.net/browse/PDA-802
-      unread: flags.get(Feature.ProductionAndAlpha) ? false : isUnread,
-      id: String(threadID),
-      phoneNumber: String(number),
-      lastUpdatedAt: new Date(lastUpdatedAt * 1000),
     }
   }
