+++ packages/app/src/messages/services/index.ts
@@ -4,3 +4,4 @@
  */
 
 export * from "./message.service"
+export * from "./thread.service"
